restructure: moved FormtDate to lib folder
