+++ app/frontend/src/components/notification/NotificationCard.tsx
@@ -2,7 +2,7 @@ import { Bell, RefreshCw } from 'lucide-react';
 import Link from 'next/link';
 import type { FireNotification } from '../../types/Notifications';
 import { NotificationBadge } from './StatusBadge';
-import { FormatDate } from '../shared/FormatDate';
+import { FormatDate } from '../../lib/FormatDate';
 import { useAuth } from '../../hooks/useAuth';
 import { NotificationLink } from './NotificationLink';
